+++ src/tui/tui.go
@@ -489,11 +489,13 @@ type Renderer interface {
 	RefreshWindows(windows []Window)
 	Refresh()
 	Close()
-	PassThrough(string)
+	PassThrough(y int, x int, data string)
+	Sync(bool)
 	NeedScrollbarRedraw() bool
 
 	GetChar() Event
 
+	Top() int
 	MaxX() int
 	MaxY() int
 
@@ -525,6 +527,7 @@ type Window interface {
 	Fill(text string) FillReturn
 	CFill(fg Color, bg Color, attr Attr, text string) FillReturn
 	Erase()
+	EraseMaybe() bool
 }
 
 type FullscreenRenderer struct {